refactor(website): drop content-hub-migrate remnants, tighten schema exports [T001490]
