+++ scripts/plan-context.sh
