+++ website/src/lib/__tests__/content-publish.test.ts
@@ -7,6 +7,8 @@ const validFaq = [{ question: 'q', answer: 'a' }];
 // assert the branch/PR/auto-merge sequence and return a fixed
 // `currentSha` to drive the 409/200/422 matrix.
 function fakeGitHub({ currentSha, currentValue }: { currentSha: string; currentValue?: unknown }): GitHubClient & {
+  currentSha: string;
+  currentValue?: unknown;
   branchName?: string;
   putSha?: string;
   prNumber: number;
