+++ website/src/lib/website-db.ts
@@ -1,5 +1,5 @@
 // Types re-exported from config/types.ts for backward compatibility.
-import type { ReferenzItem, ReferenzenConfig } from '../config/types';
+import type { ReferenzenConfig } from '../config/types';
 export type { ReferenzItem, ReferenzenType, ReferenzenConfig } from '../config/types';
 
 // ── T001490: Transitional type re-exports for the admin save endpoints ──────
@@ -31,6 +31,7 @@ export type LeistungCategoryOverride = LeistungCategory;
 export type LeistungServiceOverride = LeistungServiceRow;
 export type ServiceOverride = HomepageService & {
   pageContent?: ServicePageContent;
+  hidden?: boolean;
   leistungCategoryId?: string;
   headlineKey?: string;
   headlinePrefix?: boolean;
